+++ example/lib/gallery/sections/icons.dart
@@ -465,6 +465,11 @@ class IconsSection extends StatelessWidget {
                         is16x16: true,
                       ),
                       _IconEntry(
+                        data: NesIcons16.rightArrowIndicator,
+                        label: 'rightArrowIndicator',
+                        is16x16: true,
+                      ),
+                      _IconEntry(
                         data: NesIcons16.bell,
                         label: 'bell',
                         is16x16: true,